+++ lib/rorvswild/local.rb
@@ -4,8 +4,8 @@ require "rorvswild/local/queue"
 module RorVsWild
   module Local
     def self.start(config = {})
-      Rails.application.config.middleware.unshift(RorVsWild::Local::Middleware, nil)
       RorVsWild.start(config.merge(queue: RorVsWild::Local::Queue.new))
+      Rails.application.config.middleware.unshift(RorVsWild::Local::Middleware, nil)
     end
   end
 end